pyproject.toml
[project]
name = "research-assitant-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.3.1",
    "google-adk>=1.19.0",
]
